[JENKINS-58601] (privileges management)

+++ src/main/java/hudson/plugins/jobConfigHistory/JobConfigHistoryProjectAction.java
@@ -217,7 +217,13 @@ public class JobConfigHistoryProjectAction extends JobConfigHistoryBaseAction {
 	}
 
 	@Override
+	protected void checkDeleteEntryPermission() { getAccessControlledObject().checkPermission(JobConfigHistory.DELETEENTRY_PERMISSION); }
+
+	@Override
 	public boolean hasAdminPermission() { return getAccessControlledObject().hasPermission(Jenkins.ADMINISTER); }
+
+	@Override
+	public boolean hasDeleteEntryPermission() { return getAccessControlledObject().hasPermission(JobConfigHistory.DELETEENTRY_PERMISSION);}
 
 	@Override
 	public boolean hasConfigurePermission() {
@@ -447,6 +453,7 @@ public class JobConfigHistoryProjectAction extends JobConfigHistoryBaseAction {
 	}
 
 	public final void doDeleteRevision(StaplerRequest req, StaplerResponse rsp) {
+		checkDeleteEntryPermission();
 		final String timestamp = req.getParameter("timestamp");
 		PluginUtils.getHistoryDao().deleteRevision(this.getProject(), timestamp);
 		//do nothing with the rsp
